fix: 🐛 resize error ids
