+++ apps/builder/src/page/App/components/ScaleSquare/components/WrapperContainer/index.tsx
@@ -1,28 +1,20 @@
 import { ILLA_MIXPANEL_EVENT_TYPE } from "@illa-public/mixpanel-utils"
-import { klona } from "klona/json"
 import { get } from "lodash-es"
-import { FC, MouseEvent, memo, useCallback, useContext, useMemo } from "react"
+import { FC, memo, useContext, useMemo } from "react"
 import { useTranslation } from "react-i18next"
-import { useDispatch, useSelector } from "react-redux"
+import { useSelector } from "react-redux"
 import { DropList, DropListItem, Dropdown } from "@illa-design/react"
-import { useMouseHover } from "@/page/App/components/ScaleSquare/utils/useMouseHover"
 import {
   getIsILLAEditMode,
   getSelectedComponentDisplayNames,
 } from "@/redux/config/configSelector"
-import { configActions } from "@/redux/config/configSlice"
-import { getComponentDisplayNameMapDepth } from "@/redux/currentApp/components/componentsSelector"
 import {
   getExecutionError,
   getExecutionResult,
-  getExecutionWidgetLayoutInfo,
-  getIsResizing,
 } from "@/redux/currentApp/executionTree/executionSelector"
 import { CopyManager } from "@/utils/copyManager"
-import { FocusManager } from "@/utils/focusManager"
 import { trackInEditor } from "@/utils/mixpanelHelper"
 import { ShortCutContext } from "@/utils/shortcut/shortcutProvider"
-import { isMAC } from "@/utils/userAgent"
 import { WrapperContainerProps } from "./interface"
 import { applyWrapperPendingStyle, hoverHotSpotStyle } from "./style"
 
@@ -34,16 +26,11 @@ const WrapperContainer: FC<WrapperContainerProps> = (props) => {
     widgetType,
     children,
   } = props
-  const { handleMouseEnter, handleMouseLeave } = useMouseHover()
   const executionResult = useSelector(getExecutionResult)
   const { t } = useTranslation()
   const shortcut = useContext(ShortCutContext)
-  const widgetExecutionLayoutInfo = useSelector(getExecutionWidgetLayoutInfo)
 
-  const dispatch = useDispatch()
-  const displayNameMapDepth = useSelector(getComponentDisplayNameMapDepth)
   const selectedComponents = useSelector(getSelectedComponentDisplayNames)
-  const isResizing = useSelector(getIsResizing)
   const isEditMode = useSelector(getIsILLAEditMode)
   const errors = useSelector(getExecutionError)
 
@@ -64,100 +51,6 @@ const WrapperContainer: FC<WrapperContainerProps> = (props) => {
     isAutoLimitedMode &&
     (realProps?.dynamicMaxHeight === widgetHeight ||
       realProps?.dynamicMinHeight === widgetHeight)
-
-  const handleOnSelection = useCallback(
-    (e: MouseEvent<HTMLDivElement>) => {
-      e.stopPropagation()
-
-      if (isResizing || !isEditMode) return
-      FocusManager.switchFocus("canvas", {
-        displayName: displayName,
-        type: "component",
-        clickPosition: [],
-      })
-      trackInEditor(ILLA_MIXPANEL_EVENT_TYPE.SELECT, {
-        element: "component",
-        parameter1: "click",
-      })
-
-      if ((isMAC() && e.metaKey) || e.shiftKey || (!isMAC() && e.ctrlKey)) {
-        let currentSelectedDisplayName = klona(selectedComponents)
-        const index = currentSelectedDisplayName.findIndex(
-          (currentDisplayName) => displayName === currentDisplayName,
-        )
-        if (index !== -1) {
-          currentSelectedDisplayName.splice(index, 1)
-        } else {
-          currentSelectedDisplayName.push(displayName)
-        }
-
-        const depths = currentSelectedDisplayName.map((displayName) => {
-          return displayNameMapDepth[displayName]
-        })
-        let isEqual = depths.every((depth) => depth === depths[0])
-        if (!isEqual) {
-          return
-        }
-        if (currentSelectedDisplayName.length > 1) {
-          const firstParentNode =
-            widgetExecutionLayoutInfo[currentSelectedDisplayName[0]].parentNode
-          const isSameParentNode = currentSelectedDisplayName.every(
-            (displayName) => {
-              const parentNode =
-                widgetExecutionLayoutInfo[displayName].parentNode
-              return parentNode === firstParentNode
-            },
-          )
-          if (!isSameParentNode) {
-            const lastParentNode =
-              widgetExecutionLayoutInfo[
-                currentSelectedDisplayName[
-                  currentSelectedDisplayName.length - 1
-                ]
-              ].parentNode
-            currentSelectedDisplayName = currentSelectedDisplayName.filter(
-              (displayName) => {
-                const currentParentNode =
-                  widgetExecutionLayoutInfo[displayName].parentNode
-                return lastParentNode === currentParentNode
-              },
-            )
-          }
-        }
-        currentSelectedDisplayName = Array.from(
-          new Set(currentSelectedDisplayName),
-        )
-        dispatch(
-          configActions.updateSelectedComponent(currentSelectedDisplayName),
-        )
-        return
-      }
-
-      dispatch(configActions.updateSelectedComponent([displayName]))
-    },
-    [
-      dispatch,
-      displayName,
-      displayNameMapDepth,
-      isEditMode,
-      isResizing,
-      selectedComponents,
-      widgetExecutionLayoutInfo,
-    ],
-  )
-
-  const handleContextMenu = useCallback(
-    (e: MouseEvent<HTMLDivElement>) => {
-      FocusManager.switchFocus("canvas", {
-        displayName: displayName,
-        type: "component",
-        clickPosition: [],
-      })
-      e.stopPropagation()
-      dispatch(configActions.updateSelectedComponent([displayName]))
-    },
-    [displayName, dispatch],
-  )
 
   return (
     <Dropdown
@@ -199,10 +92,6 @@ const WrapperContainer: FC<WrapperContainerProps> = (props) => {
         css={hoverHotSpotStyle}
         data-displayname={displayName}
         data-parentnode={parentNodeDisplayName}
-        onMouseEnter={handleMouseEnter}
-        onMouseLeave={handleMouseLeave}
-        onClick={handleOnSelection}
-        onContextMenu={handleContextMenu}
       >
         <div
           css={applyWrapperPendingStyle({
