+++ apps/builder/src/widgetLibrary/ChatWidget/chat.tsx
@@ -72,7 +72,11 @@ export const ChatWidget: FC<ChatWidgetProps> = (props) => {
   const handleResizeStart: ResizeStartCallback = (e) => {
     e.preventDefault()
     e.stopPropagation()
-    dispatch(executionActions.setResizingNodeIDsReducer([displayName]))
+    dispatch(
+      executionActions.setResizingNodeIDsReducer([
+        `${displayName}-resize-footer`,
+      ]),
+    )
   }
 
   const handleOnResizeStop: ResizeCallback = useCallback(
